+++ scripts/linux.sh
@@ -24,13 +24,16 @@ rm -rf "${FFI_LINUX_DIR}"
 pip install cargo-zigbuild==0.19.8
 
 # Build (GLIBC 2.17)
+cargo zigbuild -p nostr-sdk-ffi --target i686-unknown-linux-gnu.2.17 --release
 cargo zigbuild -p nostr-sdk-ffi --target x86_64-unknown-linux-gnu.2.17 --release
 cargo zigbuild -p nostr-sdk-ffi --target aarch64-unknown-linux-gnu.2.17 --release
 
 # Make directories
+mkdir -p "${FFI_LINUX_DIR}/x86"
 mkdir -p "${FFI_LINUX_DIR}/x86_64"
 mkdir -p "${FFI_LINUX_DIR}/aarch64"
 
 # Copy dynamic libraries
+cp "${TARGET_DIR}/i686-unknown-linux-gnu/release/${CDYLIB}" "${FFI_LINUX_DIR}/x86"
 cp "${TARGET_DIR}/x86_64-unknown-linux-gnu/release/${CDYLIB}" "${FFI_LINUX_DIR}/x86_64"
 cp "${TARGET_DIR}/aarch64-unknown-linux-gnu/release/${CDYLIB}" "${FFI_LINUX_DIR}/aarch64"