+++ CHANGELOG.md
@@ -34,3 +34,4 @@
 ### Added
 
 * Add support for `i686-pc-windows-msvc` and `aarch64-pc-windows-msv`
+* Add support to `i686-unknown-linux-gnu`